popovers: Extract `playground_links_popover` into separate module.
